+++ R/CytoMethQC.R
@@ -12,7 +12,8 @@ plotSingleMethStats(methData[[3]])
 
 plotMethStats(methData, conf, save = T)
 plotMethStatsSummary(methData, conf, save = T)
-plotMethLevels(methData, conf, share = T, save = F)
+plotMethLevels(methData, conf, share = T, save = T)
 
 plotCpGAnnotation(methData, hypo_hyper_def = c(20,80), conf, share = T, save = T)
-  
+
+plotCntCommonCpG(methData, conf, save = T)